refactor: clarify deck list boundaries
