+++ src/pages/deck-list/ui/DeckListView.tsx
@@ -1,33 +1,15 @@
 /**
- * @file Composes the deck feature's complete Deck List Template screen.
- * Data and callbacks arrive through props, which keeps this presentation usable in both a live
- * container and Storybook.
+ * @file Renders the Deck List presentation from prepared sections and callbacks.
  */
 
 import * as React from "react";
 
-import type { Deck } from "@/entities/deck";
-import { Layout } from "@/shared/ui/layout";
+import type { DeckListItem, DeckListSections } from "../selectors/buildDeckListSections";
+import { DeckListCard, type DeckListCardActions } from "./DeckListCard";
 
-import { DeckListCard, type DeckListCardActions, type DeckListStudyProgress } from "./DeckListCard";
-
-export interface DeckListItem {
-  deck: Deck;
-  cardCount: number;
-  studyProgress?: DeckListStudyProgress;
-}
-
-export interface DeckListSections {
-  studying: DeckListItem[];
-  other: DeckListItem[];
-}
-
-export interface DeckListTemplateProps {
+export interface DeckListViewProps {
   sections: DeckListSections;
-  layout?: React.ComponentProps<typeof Layout>;
   deckCard?: DeckListCardActions;
-  feedbackSlot?: React.ReactNode;
-  dialogSlot?: React.ReactNode;
 }
 
 /**
@@ -37,9 +19,7 @@ export interface DeckListTemplateProps {
 const countLabel = (count: number) => `${count} ${count === 1 ? "deck" : "decks"}`;
 
 /**
- * Composes the complete Deck List Section screen from reusable UI components.
- * All data and callbacks arrive through props, allowing the same screen to run in containers,
- * tests, and Storybook.
+ * Renders one labeled group of Deck List items.
  */
 const DeckListSection: React.FC<{
   title: string;
@@ -76,23 +56,19 @@ const DeckListSection: React.FC<{
 };
 
 /**
- * Composes the complete Deck List Template screen from reusable UI components.
- * All data and callbacks arrive through props, allowing the same screen to run in containers,
- * tests, and Storybook.
+ * Renders the Deck List presentation from prepared sections and action callbacks.
  */
-export const DeckListView: React.FC<DeckListTemplateProps> = (props) => {
+export const DeckListView: React.FC<DeckListViewProps> = (props) => {
   const total = props.sections.studying.length + props.sections.other.length;
 
   return (
-    <Layout showHeader {...props.layout}>
-      {props.feedbackSlot}
-      {props.dialogSlot}
+    <>
       <div className="flex items-baseline justify-between gap-3">
         <h1 className="break-words text-title font-bold text-ink">Decks</h1>
         <span className="shrink-0 text-caption text-ink-muted">{countLabel(total)}</span>
       </div>
       <DeckListSection title="Studying" note="recent first" items={props.sections.studying} actions={props.deckCard} />
       <DeckListSection title="Other decks" note="A–Z" items={props.sections.other} actions={props.deckCard} />
-    </Layout>
+    </>
   );
 };
